+++ packages/mail/src/main.ts
@@ -2,15 +2,15 @@ import nodemailer from 'nodemailer';
 import { Metadata, IScanNode, ScanHook } from '@augejs/module-core';
 import Mail from 'nodemailer/lib/mailer';
 
-export const MAIL_TOKEN = 'MAIL_TOKEN';
-
 export {
   Mail
 }
 
 // https://nodemailer.com/about/
 
-const MAIL_IDENTIFIER = 'mail';
+export const ConfigName = 'mail';
+
+const MAIL_IDENTIFIER = Symbol.for(ConfigName);
 
 export function MailTransport(opts?: any): ClassDecorator {
   return function(target: Function) {
@@ -18,12 +18,12 @@ export function MailTransport(opts?: any): ClassDecorator {
       ScanHook(
         async (scanNode: IScanNode, next: Function) => {
           const config: any = {
-            ...scanNode.context.rootScanNode?.getConfig(MAIL_IDENTIFIER),
-            ...scanNode.getConfig(MAIL_IDENTIFIER),
+            ...scanNode.context.rootScanNode?.getConfig(ConfigName),
+            ...scanNode.getConfig(ConfigName),
             ...opts
           }
           const mail = nodemailer.createTransport(config);
-          scanNode.context.container.bind(MAIL_TOKEN).toConstantValue(mail);
+          scanNode.context.container.bind(MAIL_IDENTIFIER).toConstantValue(mail);
           await next();
         }
       ),