+++ CHANGELOG.md
@@ -1,4 +1,50 @@
-#  (2020-09-30)
+# [](https://github.com/augejs/standard-modules/compare/v1.0.1-rc.5...v) (2020-10-17)
+
+
+
+## [1.0.1-rc.5](https://github.com/augejs/standard-modules/compare/v1.0.1-rc.4...v1.0.1-rc.5) (2020-10-05)
+
+
+### Features
+
+* release for 1.0.0-rc.5 ([1f22978](https://github.com/augejs/standard-modules/commit/1f229782fc079e1909cd8bd7cf82e6c205ac37ac))
+
+
+
+## [1.0.1-rc.4](https://github.com/augejs/standard-modules/compare/v1.0.1-rc.3...v1.0.1-rc.4) (2020-10-05)
+
+
+### Features
+
+* release for 1.0.0-rc.4 ([0625ae4](https://github.com/augejs/standard-modules/commit/0625ae4967c7663dc9066a3276940e6bacd16707))
+
+
+
+## [1.0.1-rc.3](https://github.com/augejs/standard-modules/compare/v1.0.1-rc.2...v1.0.1-rc.3) (2020-10-05)
+
+
+### Features
+
+* release for 1.0.0-rc.3 ([29bee24](https://github.com/augejs/standard-modules/commit/29bee24cb76eb6659d9b26f1862fcdb05cc25125))
+
+
+
+## [1.0.1-rc.2](https://github.com/augejs/standard-modules/compare/v1.0.1-rc.1...v1.0.1-rc.2) (2020-10-05)
+
+
+### Bug Fixes
+
+* add support for file loader processor return void type ([27bd47b](https://github.com/augejs/standard-modules/commit/27bd47bb7a2c5421146d892f2e65d0cff63cc839))
+* bugs ([9241d7a](https://github.com/augejs/standard-modules/commit/9241d7a4cfcac453f500a446904598349cf87f3b))
+
+
+### Features
+
+* requestparams support async function ([f42cc13](https://github.com/augejs/standard-modules/commit/f42cc13361473a3e592d2063cf90c55c9dde8c9a))
+
+
+
+## 1.0.1-rc.1 (2020-09-30)